Scripts folder and don't ignore meta files
