+++ LICENSE.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 033ebc5d0bd307e489514ce520f1e6b8
+timeCreated: 1487779940
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
