+++ Scripts/CapturyAvatarManager.cs
