+++ Scripts/CapturyLeapIntegration.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ce1044914dcdf1141a01d68eed29d2df
+timeCreated: 1487779940
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
