+++ Scripts/CapturyNetworkPlugin.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 452456eac1ae78c41b91c007c23fd354
+timeCreated: 1487779940
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
